split commands into separate files
