+++ pyroll/cli/__init__.py
@@ -1,4 +1,4 @@
-from .program import State
+from .program.state import State
 from .rich import console
 
 VERSION = "2.0.0"
